修复 部分Unity版本 枚举值为string的问题

+++ solar_engine_unity_plugin/Assets/SolarEngineSDK/Util/SolarEngineUtil.cs
@@ -391,7 +391,7 @@ namespace SolarEngine
 
            attributesDict.Add(SolarEngine.Analytics.SEConstant_IAP_PCount, attributes.product_num);
 
-           attributesDict.Add(SolarEngine.Analytics.SEConstant_IAP_Paystatus, attributes.paystatus);
+           attributesDict.Add(SolarEngine.Analytics.SEConstant_IAP_Paystatus, (int)attributes.paystatus);
 
            attributesDict.Add(SolarEngine.Analytics.SEConstant_IAP_FailReason,
                attributes.fail_reason == null ? "" : attributes.fail_reason);
